% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_rhs_names.R
\name{.star_rhs_replace}
\alias{.star_rhs_replace}
\title{Helper function for star_rhs_names}
\usage{
.star_rhs_replace(star, pattern, line1, line2)
}
\arguments{
\item{star}{the stargazer output}

\item{pattern}{the regex pattern}

\item{line1}{the first line for the variable name}

\item{line2}{the second line for the variable name}
}
\value{
the update stargazer output
}
\description{
A helper function to replace the names of the stargazer variables
and possibly add a second line on the standard error.
This function works with regression only
}
