% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_columns_tex}
\alias{get_num_columns_tex}
\title{get the number of stargazer columns}
\usage{
get_num_columns_tex(star)
}
\arguments{
\item{star}{the \code{stargazer} output}
}
\description{
get the number of stargazer columns
}

