% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_star_line_numbers}
\alias{get_star_line_numbers}
\title{Get \code{stargazer} line numbers by section}
\usage{
get_star_line_numbers(star, reg = TRUE)
}
\arguments{
\item{star}{the \code{stargazer} output}

\item{reg}{if the stargazer is a regression}
}
\value{
a list with the line number for the \code{start.lines},
\code{main.lines}, and \code{end.lines} of the \code{stargazer}
output
}
\description{
To get the start line numbers, main line numbers, and end line numbers
of \code{stargazer} output
}
