% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_star_lines}
\alias{get_star_lines}
\title{get the star lines from \code{stargazer} output}
\usage{
get_star_lines(star, reg)
}
\arguments{
\item{star}{the \code{stargazer} ouput}

\item{reg}{(logical) set to TRUE for a regression}
}
\value{
a list with the \code{start.lines},
\code{main.lines}, and \code{end.lines} of the \code{stargazer}
output
}
\description{
This function will return a list
}
