% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_lhs_names.R
\name{getmode}
\alias{getmode}
\title{Mode function}
\usage{
getmode(v)
}
\arguments{
\item{v}{the vector}
}
\value{
the mode
}
\description{
https://www.tutorialspoint.com/r/r_mean_median_mode.htm
}
