% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.latex}
\alias{is.latex}
\title{Check if stargazer output is in latex form}
\usage{
is.latex(star)
}
\arguments{
\item{star}{the \code{stargazer} output}
}
\description{
Check if stargazer output is in latex form
}

