% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.text}
\alias{is.text}
\title{Check if stargazer output is in text form}
\usage{
is.text(star)
}
\arguments{
\item{star}{the \code{stargazer} output}
}
\description{
Check if stargazer output is in text form
}
