% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_cleaning_functions.R
\name{si_col}
\alias{si_col}
\title{A function to for scientific unit columns from the latex
\code{siunitx} package}
\usage{
si_col(table.format, rep.times = 1)
}
\arguments{
\item{table.format}{the latex \code{siunitx}. For example \code{3.2}
is three integers and two decimals. \code{table.format} can be a
string or numeric.}

\item{rep.times}{the number of times to repeat the given column}
}
\description{
The latex \code{siunitx} package has to be loaded in the preamble.
See this answer on tex stackexchange:
\url{http://tex.stackexchange.com/a/2747} This function is useful
with \code{paste0} when updating column to work \code{siunitx} and
aligning decimals.
}
