% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_cleaning_functions.R
\name{star_asterisks}
\alias{star_asterisks}
\title{A function to replace asterisks with latex compatible values}
\usage{
star_asterisks(star)
}
\arguments{
\item{star}{the \code{stargazer} output}
}
\value{
updated stargazer output with latex compaitble
asterisks
}
\description{
A function to replace asterisks with latex compatible values
}
