% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_get_lhs_names.R
\name{star_get_lhs_names}
\alias{star_get_lhs_names}
\title{Extract Left-Hand Side Variable Names from a Stargazer Table}
\usage{
star_get_lhs_names(star)
}
\arguments{
\item{star}{A character vector representing the lines of a LaTeX table
generated by `stargazer`.  This is typically obtained by reading the
output of `stargazer` into R using `readLines()`.}
}
\value{
A character vector containing the names of the left-hand side
  variables.
}
\description{
This function extracts the names of the left-hand side (dependent) variables
from a `stargazer` generated LaTeX table.  It assumes the default
`stargazer` table format where the dependent variable names are listed
below the "Dependent variable" line.
}
\details{
This function relies on the specific formatting of default
  `stargazer` tables.  If you have customized the table's appearance
  (e.g., by adding or removing rows, or changing the labels), this function
  may not work correctly.  It also assumes that the table is a regression
  table and not another type of table.  It will only work if the table
  includes the line "Dependent variable".
}
\examples{
library(stargazer)

data(mtcars)
mod.mtcars.1 <- lm(mpg ~ hp + wt, mtcars)
mod.mtcars.2 <- lm(mpg ~ hp + wt + cyl, mtcars)
mod.mtcars.3 <- lm(hp ~ wt + cyl, mtcars)

star_get_lhs_names(stargazer(mod.mtcars.1))
star_get_lhs_names(stargazer(mod.mtcars.1, mod.mtcars.2))
star_get_lhs_names(stargazer(mod.mtcars.1, mod.mtcars.2, mod.mtcars.3))
}
