% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_cleaning_functions.R
\name{star_insert_row}
\alias{star_insert_row}
\title{Insert a row in a stargazer table}
\usage{
star_insert_row(star, string, insert.after)
}
\arguments{
\item{star}{the \code{stargazer} output}

\item{string}{the string of text to insert into the table}

\item{insert.after}{insert \code{string} after \code{inster.after}}
}
\value{
the updated \code{stargazer} output
}
\description{
Insert a row in a stargazer table
}

