% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_cleaning_functions.R
\name{star_ncol}
\alias{star_ncol}
\title{get the number of stargazer columns}
\usage{
star_ncol(star)
}
\arguments{
\item{star}{the \code{stargazer} output}
}
\value{
the number of columns in the \code{stargazer} output
}
\description{
get the number of stargazer columns
}
\examples{
library(stargazer)
data(mtcars)
star.out <- stargazer(mtcars)
star_ncol(star.out)
}
