% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_table_notes.R
\name{star_notes_caption}
\alias{star_notes_caption}
\title{stargazer table notes using latex \code{caption*} package}
\usage{
star_notes_caption(star, note)
}
\arguments{
\item{star}{\code{stargazer} output}

\item{note}{the note to be added to the latex table}
}
\value{
the updated \code{stargazer} output
}
\description{
Only latex is supported
}
\examples{
library(stargazer)
data(mtcars)
##First set up models without weight
mod.mtcars.1 <- lm(mpg ~ hp, mtcars)
mod.mtcars.2 <- lm(mpg ~ hp + cyl, mtcars)
star.out <- stargazer(mod.mtcars.1, mod.mtcars.2)
star.out <- star_notes_caption(star.out, "custom note")
print(star.out)
}

