% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_table_notes.R
\name{star_notes_tex}
\alias{star_notes_tex}
\title{\code{stargazer} tex table notes}
\usage{
star_notes_tex(star, note.type = c("caption", "threeparttable"), note)
}
\arguments{
\item{star}{\code{stargazer} output}

\item{note.type}{the type of latex note to be used. Options include
\code{"caption"} for use with the latex caption package and
\code{"threeparttable"} for use with the latex threeparttable
package. Only the first element of type will be used}

\item{note}{the note to be added to the latex table}
}
\description{
\code{stargazer} tex table notes
}
\examples{
## -- Regression Example -- ##
library(stargazer)
data(mtcars)
mod.mtcars.1 <- lm(mpg ~ hp + wt, mtcars)
mod.mtcars.2 <- lm(mpg ~ hp + wt + cyl, mtcars)
mod.mtcars.3 <- lm(hp ~ wt + cyl, mtcars)
##latex example
star.out <- stargazer(mod.mtcars.1, mod.mtcars.2, mod.mtcars.3,
                      type = "latex")
print(star.out)
##use latex caption package
star_notes_tex(star.out, note = "my table notes")
##use latex caption package
star_notes_tex(star.out, note.type = "caption", note = "my table notes")
##use latex threeparttable package
star_notes_tex(star.out, note.type = "threeparttable", note = "my table notes")
}
