% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_panel.R
\name{star_panel}
\alias{star_panel}
\title{To get a panel of stargazer output}
\usage{
star_panel(..., starlist = NULL, panel.names, reg = TRUE,
  same.lhs.vars = TRUE, same.summary.stats = TRUE,
  panel.label.fontface = "plain")
}
\arguments{
\item{...}{\code{stargazer} output}

\item{starlist}{a list of \code{stargazer} output}

\item{panel.names}{the names for the panels in the same order as
the stargazer output}

\item{reg}{(logical) set to TRUE if the panels are regression
output; defualt is TRUE}

\item{same.lhs.vars}{(logical) set to \code{TRUE} if all panels
have the same left-hand-side variables. Currently, only
\code{TRUE} is supported}

\item{same.summary.stats}{(logical) set to \code{TRUE} if all
panels have the same summary stats. Default is \code{TRUE}. If
set to \code{FALSE}, different panels can have different
summary stats. \code{same.summary.stats = FALSE} is only
supported for regression and so \code{reg} must be set to
\code{TRUE}}

\item{panel.label.fontface}{The fontface for the panel
labels. Options are "plain", "italic", "bold",
"bold.italic". Default is "plain".}
}
\value{
the updated \code{stargazer} output
}
\description{
Currently only the same lhs variables and
the same summary statistics across all panels are allowed
}
\examples{
## -- Regressoin example -- ##
library(stargazer)
data(mtcars)
##First set up models without weight
mod.mtcars.1 <- lm(mpg ~ hp, mtcars)
mod.mtcars.2 <- lm(mpg ~ hp + cyl, mtcars)
star.out.1 <- stargazer(mod.mtcars.1, mod.mtcars.2, keep.stat = "n")
##Second set of models with weight as a regressor
mod.mtcars.3 <- lm(mpg ~ hp + wt, mtcars)
mod.mtcars.4 <- lm(mpg ~ hp + cyl + wt, mtcars)
star.out.2 <- stargazer(mod.mtcars.1, mod.mtcars.2, keep.stat = c("n", "rsq"))

##stargazer panel -- same summary statistics across panels.
star.panel.out <- star_panel(star.out.1, star.out.2,
                             panel.names = c("Without Weight", "With Weight")
                            )
print(star.panel.out)
##write to a tex file and compile to check output
\dontrun{
tex_write(star.panel.out, file = "my_tex_file.tex", headers = TRUE)
}

##stargazer panel -- different summary statistics across panels.
star.panel.out2 <- star_panel(star.out.1, star.out.2,
                              same.summary.stats = FALSE,
                              panel.names = c("Without Weight", "With Weight")
                             )
print(star.panel.out2)
##write to a tex file and compile to check output
\dontrun{
tex_write(star.panel.out2, file = "my_tex_file2.tex", headers = TRUE)
}
}
