% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_tex_write.R
\name{star_tex_end}
\alias{star_tex_end}
\title{Create the end of a tex document}
\usage{
star_tex_end(file, append = TRUE)
}
\arguments{
\item{file}{a \code{character} string with the file name}

\item{append}{append to the end of a tex file; defaults to TRUE}
}
\description{
Create the end of a tex document
}
