% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_cleaning_functions.R
\name{star_tex_sidewaystable}
\alias{star_tex_sidewaystable}
\title{Turn a latex table into a \code{sidewaystable} latex table}
\usage{
star_tex_sidewaystable(star)
}
\arguments{
\item{star}{the \code{stargazer} output}
}
\value{
the latex table turned sideways
}
\description{
Turn a latex table into a \code{sidewaystable} latex table
}
\examples{
library(stargazer)
data(mtcars)
star.out <- stargazer(mtcars)
star_tex_sidewaystable(star.out)
}

