% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_tex_write.R
\name{star_tex_start}
\alias{star_tex_start}
\title{Create the start of a tex}
\usage{
star_tex_start(file)
}
\arguments{
\item{file}{a \code{string} with the file name}
}
\description{
Create the start of a tex
}
