% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_write_and_compile_pdf.R
\name{star_write_and_compile_pdf}
\alias{star_write_and_compile_pdf}
\title{Write to a file and compile to pdf}
\usage{
star_write_and_compile_pdf(..., starlist = NULL, file)
}
\arguments{
\item{...}{stargazer ouptut}

\item{starlist}{a list of stargazer output to write to a file}

\item{file}{where the file will be saved}
}
\description{
\code{star_write_and_pdf_compile} will compile the table as a pdf
and then write using \code{star_tex_write} without headers, so that
table can be directly imported into a latex file
}
\seealso{
\code{star_tex_write}
}
