% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starpolish_main.R
\docType{package}
\name{starpolish-package}
\alias{starpolish}
\alias{starpolish-package}
\title{Polish and manipulate stargazer output for latex files}
\description{
starpolish allows you to manipulate stargazer output for latex files
}

