% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starpolishr_main.R
\docType{package}
\name{starpolishr-package}
\alias{starpolishr}
\alias{starpolishr-package}
\title{Polish and manipulate stargazer output for latex files}
\description{
starpolish allows you to manipulate stargazer output for latex files
}

