% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_write.R
\name{tex_write}
\alias{tex_write}
\title{Write a stargazer output to a tex file}
\usage{
tex_write(..., starlist = NULL, file, headers = FALSE, append = FALSE)
}
\arguments{
\item{...}{stargazer ouptut}

\item{starlist}{a list of stargazer output to write to a file}

\item{file}{where the file will be saved}

\item{headers}{(logical) to include the start of a tex afile}

\item{append}{logical to append at the end of a tex file; defaults
to \code{FALSE}}
}
\description{
Write a stargazer output to a tex file
}

