% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_cleaning_functions.R
\name{star_add_column_numbers}
\alias{star_add_column_numbers}
\title{Insert column numbrers in a table}
\usage{
star_add_column_numbers(star, insert.after, skip.col.1 = TRUE)
}
\arguments{
\item{star}{the \code{stargazer} output}

\item{insert.after}{insert the column numbers after
\code{inster.after}}

\item{skip.col.1}{skip the first column? If \code{TRUE} no column
number will be placed after the first column}
}
\value{
an updated version of the stargazer table with
}
\description{
Currently this only works for latex. This function is helpful when
creating a stargazer table from matrix output as stargazer does not
automatically use column numbers for matrices
}
\examples{
library(stargazer)
data(mtcars)
star.out <- stargazer(as.matrix(head(mtcars)))
print(star.out)
star.out <- star_add_column_numbers(star.out, insert.after = 10)
print(star.out)
}

